# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Stade de Mbour",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tickets": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Liste des tickets",
                "operationId": "0082053c3590628e674a11ec0d1466e6",
                "responses": {
                    "200": {
                        "description": "Liste des tickets"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tickets/{id}": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Détails d'un ticket",
                "operationId": "62091a82e5103c19191c4c1ebd30ec6f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID du ticket",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du ticket"
                    },
                    "404": {
                        "description": "Ticket non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Mettre à jour un ticket",
                "operationId": "1010dcd7a1f352766df66431f1df4745",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID du ticket",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "prix": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2000
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "sold"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ticket mis à jour avec succès"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Ticket non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Supprimer un ticket",
                "operationId": "c37b65ae6ff3ffa8c9e0c14ea4d6a9f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID du ticket",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket supprimé avec succès"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Ticket non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/generate-tickets": {
            "post": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Générer plusieurs tickets avec QR codes",
                "operationId": "dc113fe05642dd429f636ea8190459b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ticket_category_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "039ff624-b90c-4b06-8a65-cfdd647f646c"
                                    },
                                    "event_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "89ff624-b90c-4b06-8a65-cfdd647f6422"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "prix": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1000
                                    },
                                    "channel": {
                                        "type": "string",
                                        "example": "online"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "available"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tickets générés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid",
                                                "example": "039ff624-b90c-4b06-8a65-cfdd647f646c"
                                            },
                                            "ticket_category_id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "event_id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "prix": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 1000
                                            },
                                            "qr_code": {
                                                "type": "string",
                                                "example": "https://.../storage/qrcodes/ticket_1.png"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation des données"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/purchase-ticket": {
            "post": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Achat de ticket",
                "operationId": "059263c31f7189fabaa5e7888683f146",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "event_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "039ff624-b90c-4b06-8a65-cfdd647f646c"
                                    },
                                    "category_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "949ff624-b90c-4b06-8a65-cfdd647f6411"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+221777000111"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Achat de ticket réussi"
                    },
                    "400": {
                        "description": "Pas assez de tickets disponibles"
                    },
                    "422": {
                        "description": "Erreur de validation des données"
                    }
                }
            }
        },
        "/api/get-payment-url/{saleId}": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Obtenir l'URL de paiement PayTech pour une vente",
                "operationId": "1c89a3bf8a9000e216ae37a3d9a034bb",
                "parameters": [
                    {
                        "name": "saleId",
                        "in": "path",
                        "description": "UUID de la vente (Sale)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Renvoie l'URL de paiement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "payment_url": {
                                            "type": "string",
                                            "example": "https://paytech.sn/redirect_url"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vente non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/events/{buyer_id}/tickets": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Récupérer les tickets vendus à l'utilisateur authentifié",
                "operationId": "482b2be1f3ad3e0ac00931d122496613",
                "responses": {
                    "200": {
                        "description": "Liste des tickets vendus à l'utilisateur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}